+++ tests/test_fitter.py
@@ -106,25 +106,25 @@ def test_ftest_nb():
         frozen=False,
     )
     ft = f.ftest(F2, "Spindown", remove=False)
-    assert isinstance(ft["ft"], float) or isinstance(ft["ft"], bool)
+    assert isinstance(ft["ft"], (float, bool))
     # Test return the full output
     Ftest_dict = f.ftest(F2, "Spindown", remove=False, full_output=True)
-    assert isinstance(Ftest_dict["ft"], float) or isinstance(Ftest_dict["ft"], bool)
+    assert isinstance(Ftest_dict["ft"], (float, bool))
     # Test removing parameter
     F1 = param.prefixParameter(
         parameter_type="float", name="F1", value=0.0, units=u.Hz / u.s, frozen=False
     )
     ft = f.ftest(F1, "Spindown", remove=True)
-    assert isinstance(ft["ft"], float) or isinstance(ft["ft"], bool)
+    assert isinstance(ft["ft"], (float, bool))
     Ftest_dict = f.ftest(F1, "Spindown", remove=True, full_output=True)
-    assert isinstance(Ftest_dict["ft"], float) or isinstance(Ftest_dict["ft"], bool)
+    assert isinstance(Ftest_dict["ft"], (float, bool))
 
 
 def test_ftest_wb():
     """Test for wideband fitter class F-test."""
     wb_m = tm.get_model(os.path.join(datadir, "J0023+0923_ell1_simple.par"))
     wb_t = simulation.make_fake_toas_uniform(
-        56000.0, 56001.0, 10, wb_m, freq=1400.0 * u.MHz, obs="GBT", dm=wb_m.DM.quantity
+        56000.0, 56001.0, 10, wb_m, freq=1400.0 * u.MHz, obs="GBT", wideband=True
     )
     wb_f = fitter.WidebandTOAFitter(wb_t, wb_m)
     wb_f.fit_toas()